requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
tensorflow>=2.8.0
torch>=1.11.0
torchvision>=0.12.0

# Communication and networking
fastapi>=0.75.0
uvicorn>=0.17.0
websockets>=10.0
pydantic>=1.9.0
aiohttp>=3.8.0
redis>=4.0.0
paho-mqtt>=1.6.0

# Data processing and storage
sqlalchemy>=1.4.0
influxdb-client>=1.24.0
pymongo>=4.0.0
pyzmq>=22.0.0

# Edge computing and IoT
paho-mqtt>=1.6.0
bluepy>=1.3.0
requests>=2.27.0

# Monitoring and logging
prometheus-client>=0.13.0
grafana-api>=1.0.0
loguru>=0.6.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0
schedule>=1.1.0
psutil>=5.8.0

# Testing and development
pytest>=7.0.0
pytest-asyncio>=0.18.0
black>=22.0.0
flake8>=4.0.0
